# src/main/resources/openapi/patient.yml
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "tags": [
    {
      "name": "controller",
      "description": "Controller"
    }
  ],
  "paths": {
    "/createPatient": {
      "post": {
        "tags": [
          "controller"
        ],
        "summary": "createPatient",
        "operationId": "createPatientUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "patient",
            "description": "patient",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PatientReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/deletePatient": {
      "post": {
        "tags": [
          "controller"
        ],
        "summary": "deletePatient",
        "operationId": "deletePatientUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "patient",
            "description": "patient",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PatientReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/listAllPatient": {
      "get": {
        "tags": [
          "controller"
        ],
        "summary": "listAllPatient",
        "operationId": "listAllPatientUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/PatientResp"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    }
  },
  "definitions": {
    "PatientResp": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        }
      },
      "title": "PatientResp"
    },
    "PatientReq": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        }
      },
      "title": "PatientReq"
    }
  }
}